+++ src/scanner/IntegerSize.java
@@ -4,7 +4,7 @@ package scanner;
  * Created by HoseinGhahremanzadeh on 2/28/2017.
  */
 public class IntegerSize {
-    public static final int IntSize = 32;
-    public static final int LongSize = 32;
-    public static final int LongLongSize = 64;
+    public static final long IntSize = 32L;
+    public static final long LongSize = 32L;
+    public static final long LongLongSize = 64L;
 }
