+++ src/scanner/RealNumberSize.java
@@ -4,6 +4,6 @@ package scanner;
  * Created by HoseinGhahremanzadeh on 3/2/2017.
  */
 public class RealNumberSize {
-    public static final int FloatSize = 32;
-    public static final int DoubleSize = 64;
+    public static final long FloatSize = 32L;
+    public static final long DoubleSize = 64L;
 }
